refactor: flow becomes spender model (w/o default import)
